serverAI/requirements.txt
openai==1.79.0
pymongo==4.12.1
flask==3.1.0
requests==2.32.3
flask-cors==5.0.1
python-dotenv==1.0.1